separate viewer-camera-logic from render and it's starting to work!
